# FrontEnd/ChamberAppel/src/app/apis/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Chamber d'appel",
    "description": "Ceci est une documentation d'API Swagger pour la gestion de la Chambre d'Appel de la Cour des Comptes.",
    "contact": {
      "name": "Cour des comptes",
      "url": "https://www.courdescomptes.ma"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Utilisateurs/GetAllAsync": {
      "post": {
        "tags": [
          "Utilisateurs"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DtoFiltreUtilisateurDatatableRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DtoFiltreUtilisateurDatatableRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DtoFiltreUtilisateurDatatableRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DtoUtilisateurDatatableResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DtoUtilisateurDatatableResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DtoUtilisateurDatatableResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Utilisateurs/GetByIdAsync/{id}": {
      "get": {
        "tags": [
          "Utilisateurs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DtoUtilisateurApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DtoUtilisateurApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DtoUtilisateurApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Utilisateurs/CreateAsync": {
      "post": {
        "tags": [
          "Utilisateurs"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DtoUtilisateur"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DtoUtilisateur"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DtoUtilisateur"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DtoUtilisateurApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DtoUtilisateurApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DtoUtilisateurApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Utilisateurs/UpdateAsync": {
      "put": {
        "tags": [
          "Utilisateurs"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DtoUtilisateur"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DtoUtilisateur"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DtoUtilisateur"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DtoUtilisateurApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DtoUtilisateurApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DtoUtilisateurApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Utilisateurs/DeleteAsync/{id}": {
      "delete": {
        "tags": [
          "Utilisateurs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Utilisateurs/ExporterAsync": {
      "post": {
        "tags": [
          "Utilisateurs"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DtoFiltreUtilisateur"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DtoFiltreUtilisateur"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DtoFiltreUtilisateur"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Utilisateurs/GetPermissionsAsync/{id}": {
      "get": {
        "tags": [
          "Utilisateurs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionListApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionListApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionListApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Utilisateurs/GetRolesAsync/{id}": {
      "get": {
        "tags": [
          "Utilisateurs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RoleListApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleListApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleListApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Utilisateurs/AddPermissionsAsync": {
      "post": {
        "tags": [
          "Utilisateurs"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DtoCheckedListRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DtoCheckedListRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DtoCheckedListRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Utilisateurs/AddRolesAsync": {
      "post": {
        "tags": [
          "Utilisateurs"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DtoCheckedListRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DtoCheckedListRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DtoCheckedListRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Utilisateurs/LoginAsync": {
      "post": {
        "tags": [
          "Utilisateurs"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DtoLogin"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DtoLogin"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DtoLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DtoLoginResultApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DtoLoginResultApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DtoLoginResultApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Utilisateurs/ResetPasswordAsync": {
      "post": {
        "tags": [
          "Utilisateurs"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DtoChangePassword"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DtoChangePassword"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DtoChangePassword"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanApiResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BooleanApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "boolean"
          },
          "statusCode": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "validationErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FluentValidationErrors"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DtoChangePassword": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "oldPassword": {
            "type": "string",
            "nullable": true
          },
          "newPassword": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DtoCheckedListRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "listCheckedId": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DtoFiltreUtilisateur": {
        "type": "object",
        "properties": {
          "motsCle": {
            "type": "string",
            "nullable": true
          },
          "nomComplete": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DtoFiltreUtilisateurDatatableRequest": {
        "type": "object",
        "properties": {
          "filtre": {
            "$ref": "#/components/schemas/DtoFiltreUtilisateur"
          },
          "pagination": {
            "$ref": "#/components/schemas/DtoPagination"
          }
        },
        "additionalProperties": false
      },
      "DtoLogin": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DtoLoginResult": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "nullable": true
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DtoLoginResultApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DtoLoginResult"
          },
          "statusCode": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "validationErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FluentValidationErrors"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DtoPagination": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "orderByDirection": {
            "type": "string",
            "nullable": true
          },
          "orderBy": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DtoUtilisateur": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "personnePhysiqueId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "nom": {
            "type": "string",
            "nullable": true
          },
          "prenom": {
            "type": "string",
            "nullable": true
          },
          "nomArabe": {
            "type": "string",
            "nullable": true
          },
          "prenomArabe": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "login": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "updatedBy": {
            "type": "string",
            "nullable": true
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isActive": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "DtoUtilisateurApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DtoUtilisateur"
          },
          "statusCode": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "validationErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FluentValidationErrors"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DtoUtilisateurDatatableResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DtoUtilisateur"
            },
            "nullable": true
          },
          "recordTotal": {
            "type": "integer",
            "format": "int32"
          },
          "recordFiltred": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "FluentValidationErrors": {
        "type": "object",
        "properties": {
          "propertyName": {
            "type": "string",
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "HttpStatusCode": {
        "enum": [
          100,
          101,
          102,
          103,
          200,
          201,
          202,
          203,
          204,
          205,
          206,
          207,
          208,
          226,
          300,
          301,
          302,
          303,
          304,
          305,
          306,
          307,
          308,
          400,
          401,
          402,
          403,
          404,
          405,
          406,
          407,
          408,
          409,
          410,
          411,
          412,
          413,
          414,
          415,
          416,
          417,
          421,
          422,
          423,
          424,
          426,
          428,
          429,
          431,
          451,
          500,
          501,
          502,
          503,
          504,
          505,
          506,
          507,
          508,
          510,
          511
        ],
        "type": "integer",
        "format": "int32"
      },
      "Permission": {
        "required": [
          "groupe",
          "label"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "label": {
            "maxLength": 150,
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "integer",
            "format": "int32"
          },
          "groupe": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "PermissionListApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            },
            "nullable": true
          },
          "statusCode": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "validationErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FluentValidationErrors"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Role": {
        "required": [
          "label"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "label": {
            "maxLength": 150,
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "RoleListApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            },
            "nullable": true
          },
          "statusCode": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "validationErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FluentValidationErrors"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}